B3-wordCount/vm1/requirements.txt
fastapi
uvicorn
jinja2
grpcio
grpcio-tools
python-multipart